Create C program to demonstrate zombie process
A process which has finished its execution but still has an entry in the process table to report to its parent process is known as a zombie process.

In the following code, you can see that the parent will sleep for 20 sec, so it will complete its execution after 20 sec. But, Child will finish its execution using exit() system call while its parent process has gone for sleep.

After execution the child must report to its parent, So the child process entry has to be in the process table to report to its parent even after it has finished execution.